pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fm-dicom"
version = "0.1.0"
description = "A basic wrapper for pydicom commands. Lets you edit DICOM tags, merge patients and then export/send the changes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0.0",
    "pillow>=11.2.1",
    "pydicom>=3.0.1",
    "pynetdicom>=2.0.2",
    "pyqt6>=6.9.0",
    "python-gdcm>=3.0.24",
    "pyyaml>=6.0.2",
    "typer>=0.12.3",
]

# Optional dependencies for development/packaging
[project.optional-dependencies]
dev = [
    "pyinstaller>=6.13.0",  
]
test = [
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.0.0",
    "pytest>=8.4.1",
    "pytest-qt>=4.5.0",
]

[project.scripts]
fm-dicom = "fm_dicom.main:main"
